yogcorrectionshtml030324

+++ db/migrate/20240303192637_add_price_to_basket_items.rb
@@ -1,5 +1,5 @@
 class AddPriceToBasketItems < ActiveRecord::Migration[7.1]
   def change
-    add_column :basket_items, :price, :decimal
+    add_column :basket_items, :price, :integer
   end
 end
